refactor: move worker tests into modules
